move DAGStoreWrapper to stores package.
